BUG 6487: Extra routes to IVpnLink endpoint not in FIB
 + LportTag range for InterVpnLink endpoints has changed to a range of
   [170001..270000]. Those values does not fit in a uint16 attribute

 + Also applies a bit of refactoring to reduce the duplication of code
   as much as possible

 + It also aligns the system properties names to the ones defined
   in sdnc repository

 Apart from above bug fix this commit also deals with:

 1> Fixes in metadata writing
 2> InterVpnLink Active even if VPN not exists

+++ vpnservice/vpnmanager/vpnmanager-impl/src/main/java/org/opendaylight/netvirt/vpnmanager/intervpnlink/InterVpnLinkUtil.java
@@ -121,11 +121,12 @@ public class InterVpnLinkUtil {
      */
     public static void installLPortDispatcherTableFlow(DataBroker broker, IMdsalApiManager mdsalManager,
                                                        InterVpnLink interVpnLink, List<BigInteger> dpnList,
-                                                       Uuid vpnUuidOtherEndpoint, Integer lPortTagOfOtherEndpoint) {
+                                                       Uuid vpnUuidOtherEndpoint, Long lPortTagOfOtherEndpoint) {
         long vpnId = VpnUtil.getVpnId(broker, vpnUuidOtherEndpoint.getValue());
         for ( BigInteger dpnId : dpnList ) {
             // insert into LPortDispatcher table
-            Flow lPortDispatcherFlow = buildLPortDispatcherFlow(interVpnLink.getName(), vpnId, lPortTagOfOtherEndpoint);
+            Flow lPortDispatcherFlow = buildLPortDispatcherFlow(interVpnLink.getName(), vpnId,
+                    lPortTagOfOtherEndpoint.intValue());
             mdsalManager.installFlow(dpnId, lPortDispatcherFlow);
         }
     }
@@ -139,7 +140,7 @@ public class InterVpnLinkUtil {
     * @param lportTag DataPlane identifier of the LogicalPort.
     * @return the Flow ready to be installed
     */
-   public static Flow buildLPortDispatcherFlow(String interVpnLinkName, long vpnId, Integer lportTag) {
+   public static Flow buildLPortDispatcherFlow(String interVpnLinkName, long vpnId, int lportTag) {
        LOG.info("Inter-vpn-link : buildLPortDispatcherFlow. vpnId {}   lportTag {} ", vpnId, lportTag);
        List<MatchInfo> matches = Arrays.asList(new MatchInfo(MatchFieldType.metadata,
                                                               new BigInteger[] {
